requirements.txt
tqdm>=4.47.0
munch>=2.5.0
torchvision>=0.8.1
opencv_python_headless>=4.1.1.26
requests>=2.22.0
einops>=0.3.0
chardet>=3.0.4
x_transformers>=0.8.1
torch>=1.7.0
imagesize>=1.2.0
transformers>=4.2.2
numpy>=1.19.5
Pillow>=8.1.0
PyYAML>=5.4.1
torchtext>=0.6.0
albumentations>=0.5.2
pandas>=1.0.0
git+https://github.com/rwightman/pytorch-image-models.git@ca9b078ac7b5058d6d8d3db0fa6b30916b8fa113
flask
flask-cors
